requirements.txt
# Discord Bot Dependencies

# Core dependencies
discord.py==2.3.2
aiohttp==3.9.1
aiosqlite==0.19.0

# Time and scheduling
python-dateutil==2.8.2
pytz==2023.3

# Redis for caching (optional but recommended)
redis==5.0.1

# Utilities
python-dotenv==1.0.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0